SASCA refactor: integrate graph parser and builder
