--- parse_sasca/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-edition = "2021"
-name = "parse_sasca"
-version = "0.1.0"
-
-[dependencies]
-chumsky = "0.8.0"
-ariadne = "0.1.5"
-indexmap = "1.9.1"
