--- parse_sasca/t.txt
@@ -1,13 +1,0 @@
-NC 32
-NC 32
-
- 
-PROPERTY x = a ^ b ^ !c
-
-qsdf
- 
-
-TABLE t = [0, 1, 3, 4]
-
-
-
