+++ src/scalib_ext/scalib/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod belief_propagation;
-pub(crate) mod bp_compute;
-pub mod factor_graph;
+pub mod sasca;
 pub mod lda;
 pub(crate) mod matrixmul;
 pub mod mttest;
